add other k

+++ main.py
@@ -29,22 +29,25 @@ if __name__ == "__main__":
     
     os.makedirs('embedding_result', exist_ok=True)
     
-    run.add(
-        "calculating embedding",
-        "python embedding/[[embedding]].py [[edgelist]]",
-        {'embedding': embeddings,
-        'edgelist': getFiles('input_data')},
-        allowed_return_codes=[0,124],
-    )
+    if not os.path.exists('embedding_result'):
+        run.add(
+            "calculating embedding",
+            "python embedding/[[embedding]].py [[edgelist]]",
+            {'embedding': embeddings,
+            'edgelist': getFiles('input_data')},
+            allowed_return_codes=[0,124],
+        )
     
-    if os.path.exists('embedding/verse_exe/temp'):
-        shutil.rmtree('embedding/verse_exe/temp')
-    if os.path.exists('embedding/struc2vec_exe/temp'):
-        shutil.rmtree('embedding/struc2vec_exe/temp')
+        if os.path.exists('embedding/verse_exe/temp'):
+            shutil.rmtree('embedding/verse_exe/temp')
+        if os.path.exists('embedding/struc2vec_exe/temp'):
+            shutil.rmtree('embedding/struc2vec_exe/temp')
     
     evaluations = list()
-    evaluations.append('average_error_link_prediction')
-    evaluations.append('precision_at_k_link_prediction')
+    evaluations.append('average_error_link_prediction.py')
+    evaluations.append('precision_at_k_link_prediction.py 5')
+    evaluations.append('precision_at_k_link_prediction.py 10')
+    evaluations.append('precision_at_k_link_prediction.py 15')
     
     if os.path.exists('evaluation_result'):
         shutil.rmtree('evaluation_result')
@@ -52,7 +55,7 @@ if __name__ == "__main__":
     
     run.add(
         "evaluating",
-        "python evaluation/[[evaluation]].py [[edgelist]] embedding_result/[[embedding]]/[[edgelist]]",
+        "python evaluation/[[evaluation]] [[edgelist]] embedding_result/[[embedding]]/[[edgelist]]",
         {'evaluation': evaluations,
          'embedding': embeddings,
          'edgelist': getFiles('input_data')},
